chore: added schema model migrations to databases
