+++ iot-platform-core/main.go
@@ -1,7 +1,13 @@
 package main
 
-import "fmt"
+import (
+	"fmt"
+
+	"github.com/balub/The-IoT-Project/databases"
+)
 
 func main() {
+    databases.Connect()
+    databases.Migrate()
     fmt.Println("Hello, world.")
 }
